pyproject.toml
[project]
name = "codeleap-backend-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-rest-framework>=0.1.0",
    "split-settings>=1.0.0",
    "psycopg2-binary>=2.9.9",
]
